refactor: execCommand into function
