+++ .github/actions/loki/fuselageSnap/src/utils/execCommand.js
@@ -1,0 +1,16 @@
+import { spawn } from 'child_process';
+
+export const execCommand = (command) => {
+  return new Promise((resolve, reject) => {
+    const childProcess = spawn(command, {
+      stdio: 'inherit',
+      shell: true,
+    });
+    childProcess.on('error', (error) => {
+      reject(error);
+    });
+    childProcess.on('exit', () => {
+      resolve();
+    });
+  });
+};
